feat(#970): delete duplicate core/cache_store.py, move imports to contracts
core/cache_store.py was a full duplicate of the canonical
contracts/cache_store.py (which has additional abstract method
keys_by_pattern). Deleted the duplicate, updated 9 callers to import
from nexus.contracts.cache_store, and added missing keys_by_pattern
implementations to DragonflyCacheStore and InMemoryCacheStore. Also
removed stale ruff per-file-ignores entry.
