+++ pyproject.toml
@@ -831,7 +831,6 @@ ignore = [
 "alembic/*" = ["UP007", "E402"]  # Alembic: Union syntax, import order
 "src/nexus/backends/*.py" = ["ARG002"]  # Keep unused context params for API compatibility
 "src/nexus/bricks/ipc/driver.py" = ["ARG002"]  # Backend ABC override — params required by interface
-"src/nexus/core/cache_store.py" = ["ARG002"]  # NullCacheStore — null object pattern, args required by CacheStoreABC
 "src/nexus/ipc/driver.py" = ["ARG002"]  # Backend ABC override — context params required by interface
 "src/nexus/core/ace/trajectory.py" = ["ARG002"]  # Keep unused permission param for API compatibility
 "src/nexus/core/nexus_fs.py" = ["ARG002"]  # Constructor accepts cache params for public API
