+++ src/nexus/ipc/delivery.py
@@ -40,7 +40,7 @@ from nexus.ipc.storage.protocol import IPCStorageDriver
 from nexus.storage.zone_settings import SigningMode
 
 if TYPE_CHECKING:
-    from nexus.core.cache_store import CacheStoreABC
+    from nexus.contracts.cache_store import CacheStoreABC
     from nexus.ipc.signing import MessageSigner, MessageVerifier
 
 logger = logging.getLogger(__name__)
