+++ src/nexus/services/oauth_service.py
@@ -24,7 +24,7 @@ from nexus.core.rpc_decorator import rpc_expose
 logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
-    from nexus.core.cache_store import CacheStoreABC
+    from nexus.contracts.cache_store import CacheStoreABC
     from nexus.core.permissions import OperationContext
